+++ HeroesPowerPlant/LayoutEditor/SetObjects/ObjectClassesHeroes/List00/Object0024_Cage.cs
@@ -12,7 +12,7 @@
 
         public CageType Type
         {
-            get => (CageType)ReadLong(4);
+            get => (CageType)ReadInt(4);
             set => Write(4, (int)value);
         }
