+++ HeroesPowerPlant/LayoutEditor/SetObjects/ObjectClassesHeroes/List11/Object1101_CastleDoor.cs
@@ -4,7 +4,7 @@
     {
         public bool IsUpsideDown
         {
-            get => ReadLong(4) != 0;
+            get => ReadInt(4) != 0;
             set => Write(4, value ? 1 : 0);
         }
     }
