+++ HeroesPowerPlant/LayoutEditor/SetObjects/ObjectClassesHeroes/List11/Object1181_Celestial.cs
@@ -18,7 +18,7 @@ namespace HeroesPowerPlant.LayoutEditor
 
         public int Type
         {
-            get => ReadLong(4);
+            get => ReadInt(4);
             set => Write(4, value);
         }
