+++ HeroesPowerPlant/LayoutEditor/SetObjects/ObjectClassesHeroes/List15/Object15C0_EggBishop.cs
@@ -35,7 +35,7 @@ namespace HeroesPowerPlant.LayoutEditor
 
         public int AttackInterval
         {
-            get => ReadLong(20);
+            get => ReadInt(20);
             set => Write(20, value);
         }
