+++ HeroesPowerPlant/LayoutEditor/SetObjects/ObjectClassesHeroes/ListOther/Object_L1Type.cs
@@ -4,7 +4,7 @@
     {
         public int Type
         {
-            get => ReadLong(4);
+            get => ReadInt(4);
             set => Write(4, value);
         }
     }
